# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'New India Export / VIRASTRA API'
  version: 2.3.0
  description: "Production Express API for customer onboarding, OTP auth, KYC, cases, Razorpay payments, GST invoices, events, and admin ops.\n\n**Auth:** Email OTP only. Send `Authorization: Bearer <JWT>` for protected routes.\n\n**Money:** Amounts in INR rupees in API/DB; converted to paise only for Razorpay.\n\n**Envelope:** `{ success, data?, message? }` on success; `{ success: false, message, code?, details? }` on errors.\n\nLegacy vault/shipment routes under `/api/cases/...` exist only when `ENABLE_LEGACY_SHIPMENT=true`."
  contact:
    name: 'New India Export'
    email: support@newindiaexport.com
servers:
  -
    url: 'http://127.0.0.1:5001'
    description: Local
  -
    url: /
    description: 'Current host'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from OTP verify / login. Claims: sub, email, role, status.'
...
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: array
          items: {}
...
    EventUpsert:
      type: object
      required:
        - title
      properties:
        id:
          type: string
        title:
          type: string
        date:
          type: string
        city:
          type: string
        img:
          type: string
        seats:
          type: string
        capacity:
          type: string
        desc:
          type: string
        priceInr:
          type: number
        price:
          type: number
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        city:
          type: string
        img:
          type: string
        seats:
          type: string
        capacity:
          type: string
        desc:
          type: string
        priceInr:
          type: number
        price:
          type: number
    BrochureUpsert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        fileUrl:
          type: string
        file:
          type: string
          format: binary
...
  responses:
    BadRequest:
      description: 'Validation / bad request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 'Missing or invalid JWT'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Insufficient role / ACL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/events:
    get:
      summary: 'List events'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
    post:
      summary: 'Create / upsert event (admin)'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpsert'
  '/api/events/{eventId}':
    put:
      summary: 'Update event (admin)'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: eventId
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
    delete:
      summary: 'Soft-delete event (admin)'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: eventId
          in: path
          required: true
          schema:
            type: string
          description: null
  '/api/events/{eventId}/image':
    post:
      summary: 'Upload event image (admin)'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: eventId
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
  '/api/events/{id}/register':
    post:
      summary: 'Register for free event (402 if paid)'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
    delete:
      summary: 'Cancel event registration'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
          description: null
  /api/me/event-registrations:
    get:
      summary: 'My event registrations'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/events/registrations/me:
    get:
      summary: 'My event registrations (alias)'
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/brochures:
    get:
      summary: 'List brochures'
      tags:
        - Brochures
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
    post:
      summary: 'Create brochure (admin/ops)'
      tags:
        - Brochures
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrochureUpsert'
          application/json:
            schema:
              $ref: '#/components/schemas/BrochureUpsert'
  '/api/brochures/{id}':
    put:
      summary: 'Update brochure metadata'
      tags:
        - Brochures
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrochureUpsert'
    delete:
      summary: 'Soft-delete brochure'
      tags:
        - Brochures
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
          description: null
  '/api/brochures/{id}/file':
    get:
      summary: 'Stream brochure file'
      tags:
        - Brochures
      responses:
        '200':
          description: 'File stream'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
        - {}
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
          description: null
...
  '/api/files/{fileId}/download':
    get:
      summary: 'Download Drive-backed file (authz proxy)'
      tags:
        - Files
      responses:
        '200':
          description: 'File stream'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: fileId
          in: path
          required: true
          schema:
            type: string
          description: null
...